requirements.txt
# Core dependencies
python-dotenv>=1.1.1
aiohttp>=3.8.0
httpx>=0.28.1
requests>=2.32.5

# ESO Logs API client
git+https://github.com/knowlen/esologs-python.git

# Web scraping for ability bars
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Static site generation
jinja2>=3.1.0
markdown>=3.5.0

# Data processing
pydantic>=2.0.0

# Image generation for social media previews
Pillow>=10.0.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=24.0.0
